calculate disk space on the tmp folder Close #16
Some raspberry pi installation have a really small storage space which is not enough when trying to download ~2GB file. To overcome this problem, the tmp  tmp folder inside the addon folder, can be mounted to a external disk storage [with a ln soflink]. Therefore the disk space calculation should be based on the tmp folder.