Changed hwid to one with smaller image file size
My config: Raspbian Stretch lite, Raspberry Pi 3B+, LXDE, Kodi 18.0.1

The Widevine installer was failing to mount and extract the recovery image, so I performed a little experiment. I downloaded the SPRING recovery image myself, and tried to convert it from a binary to an ISO image and mount it, since if that worked, it would have been possible to extract libwidevine.so myself. bchunk failed because the file size was too large, which means that most likely, the installer was failing for the same reason. So, knowing I needed a smaller image, I looked up which other 32-bit ARM Chromebooks there are. It turns out there's an ARM HP Chromebook, hwid "SKATE," whose recovery image is just under 600mb. I tried to convert and mount that, and it worked. The installer still fails before completion because it can't create the widevine-config.json file, which can be fixed by converting Google's recovery.conf file to json and copying it into the right folders manually, but I can confirm that the installer is now able to extract a working libwidevine.so from the image before it fails. This shouldn't negatively affect system configs on which the SPRING recovery image already extracts properly, if anything the download will finish more quickly because of the lower file size.