ModProbe needs to run as shell on some systems
I'm using buildroot, and modprobe doesn't run unless the check_output command is run using shell=True. This pull request adds an additional optional parmater to the run_cmd definition that tells it whether to run as shell, and then passes that parameter as true when calling modprobe for the checkloop method.