Rename 'About' into 'IS Info'
The reason to rename the 'About' dialog to 'InputStream Helper
information' is that it is clearer to the user where the information is
coming from exactly.

As we expect developers to use this from their own add-on for
troubleshooting purposes, it is important to make clear what it is about
(pun intended).

This PR includes:
- Renaming About dialog in Info dialog
- Dutch translations
- Localize '(disabled)' string
- Restructure translation strings (30800: information, 30900: settings)
- Prepare translation to other languages
- Small visual improvements