Log when we download a new ChromeOS ARM image
This is crucial for our ARM users to understand a new Widevine was installed
from a specific ChromeOS image.

Having this will make it much easier for users to idenfity the cause of
their recent problems, as well as help us identify the affected ChromeOS
image in case of issues.